requirements.txt
azureml-sdk==1.48.0
azureml-core==1.48.0
azureml-dataset-runtime==1.48.0
azureml-train-core==1.48.0
scikit-learn==1.0.2
numpy==1.21.0
pandas==1.3.5
matplotlib==3.5.1
joblib==1.1.0
torch==1.13.0
